move wait_until_online to upgrader.py
