+++ services/addons/images/firmware_manager/commsignia_upgrader.py
@@ -110,20 +110,6 @@ class CommsigniaUpgrader(upgrader.UpgraderAbstractClass):
             logging.info(f"Post upgrade script executed successfully for rsu: {self.rsu_ip}.")
         except Exception as err:
             logging.error(f"Failed to execute post upgrade script for rsu {self.rsu_ip}: {err}")
-            
-    def wait_until_online(self):
-        iter = 0
-        # Ping once every second for 3 minutes until online
-        while iter < 180:
-            time.sleep(1)
-            code = subprocess.run(
-                ["ping", "-n", "-c1", self.rsu_ip], capture_output=True
-            ).returncode
-            if code == 0:
-                return 0
-            iter += 1
-        # 3 minutes pass with no response
-        return -1
 
 # sys.argv[1] - JSON string with the following key-values:
 # - ipv4_address
