+++ services/addons/images/firmware_manager/upgrader.py
@@ -1,5 +1,7 @@
 from pathlib import Path
 import abc
+import subprocess
+import time
 import download_blob
 import logging
 import os
@@ -55,6 +57,20 @@ class UpgraderAbstractClass(abc.ABC):
             logging.error(
                 f"Failed to connect to the Firmware Manager API for '{self.rsu_ip}': {err}"
             )
+
+    def wait_until_online(self):
+        iter = 0
+        # Ping once every second for 3 minutes until online
+        while iter < 180:
+            time.sleep(1)
+            code = subprocess.run(
+                ["ping", "-n", "-c1", self.rsu_ip], capture_output=True
+            ).returncode
+            if code == 0:
+                return 0
+            iter += 1
+        # 3 minutes pass with no response
+        return -1
 
     # This needs to be defined for each implementation
     @abc.abstractclassmethod
