+++ services/addons/tests/firmware_manager/test_upgrader.py
@@ -1,4 +1,4 @@
-from unittest.mock import patch
+from unittest.mock import MagicMock, patch
 import os
 
 from addons.images.firmware_manager import upgrader
@@ -142,3 +142,30 @@ def test_notify_firmware_manager_exception(mock_requests, mock_logging):
     mock_logging.error.assert_called_with(
         "Failed to connect to the Firmware Manager API for '8.8.8.8': Exception occurred during upgrade"
     )
+
+@patch("addons.images.firmware_manager.upgrader.time")
+@patch("addons.images.firmware_manager.upgrader.subprocess")
+def test_upgrader_wait_until_online_success(mock_subprocess, mock_time):
+    run_response_obj = MagicMock()
+    run_response_obj.returncode = 0
+    mock_subprocess.run.return_value = run_response_obj
+
+    test_upgrader = TestUpgrader(test_upgrade_info)
+    code = test_upgrader.wait_until_online()
+
+    assert code == 0
+    assert mock_time.sleep.call_count == 1
+
+
+@patch("addons.images.firmware_manager.upgrader.time")
+@patch("addons.images.firmware_manager.upgrader.subprocess")
+def test_upgrader_wait_until_online_timeout(mock_subprocess, mock_time):
+    run_response_obj = MagicMock()
+    run_response_obj.returncode = 1
+    mock_subprocess.run.return_value = run_response_obj
+
+    test_upgrader = TestUpgrader(test_upgrade_info)
+    code = test_upgrader.wait_until_online()
+
+    assert code == -1
+    assert mock_time.sleep.call_count == 180
